ns: Improve uplink queue batching logic
